FSW GUI countbox removed, added bw ratio adjustment
FSW GUI countbox removed, doesnt' seem to be working with FSW43. The averaging is different from RSA, that it only average for a given number of traces, or not averaging at all.  Added bw ratio adjustment for the auto bw option. There seems to be a bug with the single acquisition button. If the single measurement takes too long (averaging many times), the VISA will time out when waiting for the operation to complete.