Tek MSO added spectrum view control, fixed trace aquisition bug
Added the spectrum view panel to each channel,
allow normal and average mode in SV.
Can't find the reference level of the spectrum view.

Added spectrum trace capture, in either normal or average mode,
however, it is still bugged and the MSO doesn't respond to the command.

Added the the Average number of the Average aquisition mode.

Fixed the channel enable bug. The solution is to disable channel after addLinks,
so that the links can be correctly added.

Fixed the trace reading bug of the scope trace.
The solution is to add a fread() command to read off the terminating characters after the trace is read.
This used to stop the MSO from responding the following commands after the trace is read.