MSO54 tested in lab
Reading the curves of the scope has a problem when using the VISA communciation. The problem seems to be the synchronous readinng of the data, that the termination characters from the MSO can not be read properly, so it just stops there until TIMEOUT. Also under VISA the BytesAvailable is always 0.
When using just TCPIP communication, this is not a problem because the data is read asyncronously. So the binblockread will first just read the data, then the termination character is read out by another fread.
So far don't know how to correctly implement using VISA communication. Now it only works well with TCPIP.
The spectrum view data is still bugged. The data is there but there is no way to correctly read it out. The Tek Tech support also does not know how to read it out. They suggested to save the data on the MSO first, then read that data. It's a work around to be later implemented.